simplify axum desktop so it doesnt count as a lib
