+++ packages/fullstack/examples/axum-desktop/Cargo.toml
@@ -4,8 +4,6 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-[lib]
-
 [dependencies]
 dioxus = { workspace = true, features = ["launch", "fullstack"] }
 axum = { workspace = true, optional = true }
@@ -16,13 +14,3 @@ serde = "1.0.159"
 default = []
 server = ["axum", "tokio", "dioxus/axum"]
 desktop = ["dioxus/desktop"]
-
-[[bin]]
-name = "client"
-path = "src/client.rs"
-required-features = ["desktop"]
-
-[[bin]]
-name = "server"
-path = "src/server.rs"
-required-features = ["server"]
